split Commands_Handler_Address_Book
